+++ validator/impl/collator.cpp
@@ -5973,7 +5973,7 @@ bool Collator::create_block_candidate() {
   stats_.estimated_bytes = block_limit_status_->estimate_block_size();
   stats_.gas = block_limit_status_->gas_used;
   stats_.lt_delta = block_limit_status_->cur_lt - block_limit_status_->limits.start_lt;
-  stats_.estimated_collated_data_bytes = block_limit_status_->collated_data_stat.estimate_proof_size();
+  stats_.estimated_collated_data_bytes = (uint32_t) block_limit_status_->collated_data_stat.estimate_proof_size();
   stats_.cat_bytes = block_limit_status_->limits.classify_size(stats_.estimated_bytes);
   stats_.cat_gas = block_limit_status_->limits.classify_gas(stats_.gas);
   stats_.cat_lt_delta = block_limit_status_->limits.classify_lt(block_limit_status_->cur_lt);
